+++ esm/apprun-html.js
@@ -1,4 +1,5 @@
 import app from './apprun';
+import { safeGlobalAssign } from './type-utils';
 export { app, Component, on, update, event, customElement, ROUTER_404_EVENT, ROUTER_EVENT, safeHTML } from './apprun';
 import { createElement, render, Fragment, html, svg, run } from './vdom-lit-html';
 export { html, svg, render, run };
@@ -7,9 +8,12 @@ app.render = render;
 app.Fragment = Fragment;
 export default app;
 if (typeof window === 'object') {
-    window['React'] = window['_React'] || app;
-    window['html'] = html;
-    window['svg'] = svg;
-    window['run'] = run;
+    const globalWindow = window;
+    safeGlobalAssign(globalWindow, {
+        'React': globalWindow['_React'] || app,
+        'html': html,
+        'svg': svg,
+        'run': run
+    });
 }
 //# sourceMappingURL=apprun-html.js.map